aggiunti fix per la versione >= 2.4.2 di Magento
