+++ composer.json
@@ -8,7 +8,7 @@
   ],
   "require": {
     "php": "~7.1",
-    "magento/magento2-base": "<2.4.2"
+    "magento/magento2-base": ">=2.4.2"
   },
   "authors": [
     {
